test(evals): Split eval entrypoints by layer
Rename the generic YAML eval suite to e2e and the verifier entrypoint to verify.

Keep the full-pipeline, security-review, and verifier-only layers distinct.

Refs getsentry/getsentry#20066
